Refactor Monte Carlo operations into subpackages
Split the monolithic monte_carlo operations and config into a structured operations package (core, regressions, tests, transforms) and added explicit step factories in each group. Removed the old config.py and operations.py and introduced operations/* modules (builtins and ops) that implement and expose the per-step factories (e.g. simulation_step, raw_data_step, reference_filter_step, regression_step, etc.).

Also moved custom_op into monte_carlo, added a @custom_operation decorator and made the AST validator accept that single decorator, and added a step_type field to MCStep to record catalog-kind (helps serializing/live pipelines back to PipelineSpec). Updated package public APIs and catalog imports to reference the new modules; tests were adjusted accordingly.
