+++ SymbolicDSGE/__init__.py
@@ -7,7 +7,7 @@ from .core import (
 )
 from .kalman import KalmanFilter
 from .estimation import Estimator
-from .bundle import BundleBuilder, LoadedBundle, build_from as load_bundle
+from .bundle import BundleBuilder, build_from as load_bundle
 from . import utils
 
 __all__ = [
@@ -21,6 +21,5 @@ __all__ = [
     "linearize_model",
     # .sdsge bundle API
     "BundleBuilder",
-    "LoadedBundle",
     "load_bundle",
 ]
