+++ SymbolicDSGE/monte_carlo/reference_constructs.py
@@ -6,7 +6,7 @@ from numpy import float64, ndarray
 
 from ..core.solved_model import SolvedModel
 from .mc_constructs import DataGenReturn, NDF, ShockMapping
-from .operations import simulate_dgp
+from .operations.core.ops import simulate_dgp
 
 
 class MCReferenceConstruct:
